requirements.txt
flask==2.1.0
werkzeug==2.1.0
python-dotenv==0.19.0
requests==2.26.0
pillow==9.3.0  # Versión compatible con Python 3.11
